issue #2559 - refactored VisualProjectEstimateSection to ProjectEstimationSection
