+++ src/projects/detail/components/ProjectEstimationSection/ProjectEstimationSection.scss
@@ -1,7 +1,7 @@
 @import '~tc-ui/src/styles/tc-includes';
 
 :global {
-  .visual-design-project-estimate-section {
+  .project-estimation-section {
     display: flex;
     width: calc(100% - 40px);
     margin-left: 20px;
