+++ src/projects/detail/components/ProjectEstimationSection/index.js
@@ -1,0 +1,2 @@
+import ProjectEstimationSection from './ProjectEstimationSection'
+export default ProjectEstimationSection
