+++ src/projects/detail/components/TwoColsLayout/TwoColsLayout.scss
@@ -15,7 +15,6 @@
   padding-bottom: 10 * $base-unit;
   position: relative;
   width: 360px;
-  z-index: 1;
 
   .no-secondary-toolbar & {
     min-height: calc(100vh - 60px); /* minus primary toolbar height only */
